finally got styling where I wanted it to be and look like and cleaned up code
